requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
pydantic==1.10.11
python-dotenv==1.0.0
requests==2.31.0
databases==0.7.2
sqlalchemy==1.4.57
